Mini project snake game in c
Mini project in C many Students But Snake Game is the mini project of the first-semester built by my friend which is very different. It is a console application without a graphic library that's why it is more interesting. This game is perfect without any errors and better user interface. It is compiled in DEV C++  using the c,c++ languages . Here goes a source code you can copy And Paste in it your desired Editor of C++